Update Ban command to use DB set log channel

+++ commands/moderation/ban.js
@@ -8,7 +8,26 @@ module.exports = {
 	description: 'bans the member',
 	usage: '<id | mention>',
 	run: async (client, message, args) => {
-		const logChannel = message.guild.channels.find(c => c.name === 'logs') || message.channel;
+		// No author permissions
+		if (!message.member.hasPermission('BAN_MEMBERS')) {
+			return message.reply('❌ You do not have permissions to ban members. Please contact a staff member').then(m => m.delete(client.config.liveTime));
+		}
+		// No bot permissions
+		if (!message.guild.me.hasPermission('BAN_MEMBERS')) {
+			return message.reply('❌ I do not have permissions to ban members. Please contact a staff member').then(m => m.delete(client.config.liveTime));
+		}
+
+		let logChannel;
+
+		if (client.foundGuild.logChannels.modLog) logChannel = client.foundGuild.logChannels.modLog;
+
+		if (!logChannel)
+			return message.channel.send(
+				new RichEmbed()
+					.setColor(client.config.color.error)
+					.setTitle('Ban a User')
+					.setDescription(stripIndents`Banning a user is disabled in ${message.guild.name} (${message.guild.id}). Please contact an Administrator to setup a mod log channel.`)
+			);
 
 		if (message.deletable) message.delete();
 
@@ -20,15 +39,6 @@ module.exports = {
 		// No reason
 		if (!args[1]) {
 			return message.reply('Please provide a reason to ban.').then(m => m.delete(client.config.liveTime));
-		}
-
-		// No author permissions
-		if (!message.member.hasPermission('BAN_MEMBERS')) {
-			return message.reply('❌ You do not have permissions to ban members. Please contact a staff member').then(m => m.delete(client.config.liveTime));
-		}
-		// No bot permissions
-		if (!message.guild.me.hasPermission('BAN_MEMBERS')) {
-			return message.reply('❌ I do not have permissions to ban members. Please contact a staff member').then(m => m.delete(client.config.liveTime));
 		}
 
 		const toBan = message.mentions.members.first() || message.guild.members.get(args[0]);
@@ -74,7 +84,7 @@ module.exports = {
 					if (err) return message.channel.send(`Ban Failed: ${err}`);
 				});
 
-				logChannel.send(embed);
+				client.channels.get(logChannel).send(embed);
 			} else if (emoji === '❌') {
 				msg.delete();
